+++ proteus-cli/Cargo.toml
@@ -18,7 +18,7 @@ log = "0.4.20"
 rand = "0.8.5"
 ratatui = "0.26.3"
 symphonia = "0.5.3"
-proteus-lib = { path = "../proteus-lib", version = "0.3.1" }
+proteus-lib = { path = "../proteus-lib", version = "0.4.0" }
 colored = "3.1.1"
 indicatif = "0.18.3"
 text-to-ascii-art = "0.1.10"
