+++ proteus-lib/src/track/mod.rs
@@ -1,0 +1,9 @@
+//! Track decoding and buffering helpers.
+
+mod buffer;
+mod container;
+mod convert;
+mod single;
+
+pub use container::{buffer_container_tracks, ContainerTrackArgs};
+pub use single::{buffer_track, TrackArgs};
